fix(config.xml): white screen after start
This problem is not new. Many users have and had a problem with the splashscreen and just white screen after the app starts.

There are two problems.

1. The Splash screen delay is too short so the splash screen.hide() method in the app.component.ts file is useless. The app won't start in just 300ms! 
2. Because the fade animation is a bit too fast, the will still be a short white screen.


My workaround for this problem is to set the delay to a high-value like 10.000ms and disable the autohide value. I also increased the fade out duration to 1000ms.

Let me know your thoughts about this change. maybe there is a better solution but we should increase the delay value.

Related posts: 
https://forum.ionicframework.com/t/after-splash-screen-display-white-screen-long-time/80162
https://forum.ionicframework.com/t/ionic-2-startup-too-slow/78591/11